+++ Unit-02/01-blueprints/project/__init__.py
@@ -1,0 +1,24 @@
+from flask import Flask, url_for, redirect
+from flask_sqlalchemy import SQLAlchemy
+from flask_modus import Modus
+import os
+
+
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'postgres://localhost/users-app-db'
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
+db = SQLAlchemy(app)
+modus = Modus(app)
+
+
+from project.users.views import users_blueprint
+from project.messages.views import messages_blueprint
+
+app.register_blueprint(users_blueprint, url_prefix='/users')
+app.register_blueprint(messages_blueprint, url_prefix='/users/<int:user_id>/messages')
+
+@app.route('/')
+def root():
+	return redirect(url_for('users.index'))
+
