+++ Unit-02/01-blueprints/project/messages/templates/messages/edit.html
@@ -2,11 +2,11 @@
 
 {% block content %}
 <h1>Edit message</h1>
-<form method='POST' action="{{ url_for('messages_show', user_id=message.user_id, id=message.id) }}?_method=PATCH">
+<form method='POST' action="{{ url_for('messages.show', user_id=message.user_id, id=message.id) }}?_method=PATCH">
 	{{ form.hidden_tag() }}
 	{% for field in form if field.widget.input_type != 'hidden' %}
 		<p>
-			{{field.label }}
+			{{ field.label }}
 			{{ field }}
 			<span>
 				{% if field.errors %}
