+++ Unit-02/01-blueprints/project/messages/templates/messages/index.html
@@ -1,14 +1,14 @@
 {% extends 'base.html' %}
 
 {% block content %}
-	<a href="{{ url_for('messages_new', user_id=user.id) }}">Add a new message, {{ user.first_name }}!</a> 
+	<a href="{{ url_for('messages.new', user_id=user.id) }}">Add a new message, {{ user.first_name }}!</a> 
 	<h1>See all the messages for {{user.first_name}} </h1>
 	{% for message in user.messages %}
 	<p>
 		{{ message.content }}
 		<br>
-		<a href="{{ url_for('messages_edit',user_id=user.id, id=message.id) }}">Edit message</a>
-		<form method='POST' action="{{ url_for('messages_show', user_id=user.id, id=message.id) }}?_method=DELETE">
+		<a href="{{ url_for('messages.edit',user_id=user.id, id=message.id) }}">Edit message</a>
+		<form method='POST' action="{{ url_for('messages.show', user_id=user.id, id=message.id) }}?_method=DELETE">
 			{{ delete_form.hidden_tag() }}
 			<button>X</button>
 		</form>
