+++ Unit-02/01-blueprints/project/messages/templates/messages/new.html
@@ -2,7 +2,7 @@
 
 {% block content %}
 <h1>Add a new message</h1>
-<form method='POST' action="{{ url_for('messages_index', user_id=user.id) }}">
+<form method='POST' action="{{ url_for('messages.index', user_id=user.id) }}">
 	{{ form.hidden_tag() }}
 	{% for field in form if field.widget.input_type != 'hidden' %}
 		<p>
