+++ Unit-02/01-blueprints/project/messages/templates/messages/show.html
@@ -1,0 +1,12 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<h1>Here is your message: {{message.content}}</h1>
+
+<a href="{{ url_for('messages.edit', id=user.id) }}">Edit this message!</a>
+
+<h2>Delete user</h2>
+<form action="{{ url_for('messages.show', id=user.id) }}?_method=DELETE" method="POST">
+	<input type="submit" value="X">
+</form>
+{% endblock %}
