+++ Unit-02/01-blueprints/project/users/forms.py
@@ -5,8 +5,5 @@ class UserForm(FlaskForm):
 	first_name = StringField('First Name', [validators.DataRequired()])
 	last_name = StringField('Last Name', [validators.DataRequired()])
 
-class MessageForm(FlaskForm):
-	content = StringField('Content', [validators.DataRequired()])
-
 class DeleteForm(FlaskForm):
 	pass
