+++ Unit-02/01-blueprints/project/users/templates/users/edit.html
@@ -3,7 +3,7 @@
 {% block content %}
 <h1>Edit: {{ user.first_name }} {{ user.last_name }}</h1>
 
-<form action="{{ url_for('show', id=user.id) }}?_method=PATCH" method="POST">
+<form action="{{ url_for('users.show', id=user.id) }}?_method=PATCH" method="POST">
 	{{ form.hidden_tag() }}
 	{% for field in form  if field.widget.input_type != 'hidden' %}
 		<p>
