+++ Unit-02/01-blueprints/project/users/templates/users/index.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 
 {% block content %}
-<a href="{{ url_for('new') }}">Make a new user!</a>
+<a href="{{ url_for('users.new') }}">Make a new user!</a>
 <h1>See all users!</h1>
 
 {% for user in users %}
@@ -9,11 +9,11 @@
 <p>
     {{ user.first_name }} {{ user.last_name }}
     <br>
-    <a href="{{ url_for('messages_index', user_id=user.id) }}">See all messages for {{user.first_name}} </a>
+    <a href="{{ url_for('messages.index', user_id=user.id) }}">See all messages for {{user.first_name}} </a>
     <br>
-	<a href="{{ url_for('edit', id=user.id) }}"> Edit a user!</a>
+	<a href="{{ url_for('users.edit', id=user.id) }}"> Edit a user!</a>
 	<br>
-	<form action="{{ url_for('show', id=user.id) }}?_method=DELETE" method="POST">
+	<form action="{{ url_for('users.show', id=user.id) }}?_method=DELETE" method="POST">
 		{{delete_form.hidden_tag()}}
 		<button>Delete user</button>
 	</form>
