+++ Unit-02/01-blueprints/project/users/templates/users/new.html
@@ -2,7 +2,7 @@
 
 {% block content %}
 <h1>Add a new user!</h1>
-<form action="{{ url_for('index') }}" method='POST'>
+<form action="{{ url_for('users.index') }}" method='POST'>
 	{{ form.hidden_tag() }}
 	{% for field in form if field.widget.input_type != 'hidden' %}
 		<p>
